backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
langchain-chroma
chromadb
openai
python-dotenv
